requirements.lock.txt
click==7.1.2
databases==0.3.2
dnspython==1.16.0
dominate==2.5.1
eventlet==0.25.2
Events==0.3
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-JSGlue==0.3.1
Flask-Login==0.5.0
Flask-SocketIO==4.3.0
Flask-Sockets==0.2.1
flask-util-js==0.2.25
gevent==20.6.2
gevent-websocket==0.10.1
greenlet==0.4.16
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
monotonic==1.5
python-dateutil==2.8.1
python-engineio==3.13.0
python-socketio==4.6.0
six==1.15.0
SQLAlchemy==1.3.18
threaded==4.0.8
visitor==0.1.3
Werkzeug==1.0.1
zope.event==4.4
zope.interface==5.1.0
